+++ tests/test_web_transport.py
@@ -2,6 +2,7 @@ import gzip
 import sys
 import types
 import unittest
+from unittest import mock
 
 from flask import Flask
 
@@ -42,9 +43,17 @@ class WebTransportTests(unittest.TestCase):
             "/plugins/A_pwmenu/",
             headers={"Accept-Encoding": "gzip, deflate"},
         ):
-            response = self.plugin._html_response(self.html)
+            with mock.patch(
+                "A_pwmenu.gzip.compress",
+                wraps=gzip.compress,
+            ) as compress:
+                response = self.plugin._html_response(self.html)
 
         self.assertEqual(response.headers["Content-Encoding"], "gzip")
+        compress.assert_called_once_with(
+            self.html.encode("utf-8"),
+            compresslevel=1,
+        )
         self.assertEqual(response.headers["Vary"], "Accept-Encoding")
         self.assertEqual(gzip.decompress(response.get_data()).decode("utf-8"), self.html)
         self.assertLess(len(response.get_data()), len(self.html.encode("utf-8")) // 4)
@@ -59,6 +68,24 @@ class WebTransportTests(unittest.TestCase):
         self.assertNotIn("Content-Encoding", response.headers)
         self.assertEqual(response.mimetype, "text/html")
         self.assertEqual(response.get_data().decode("utf-8"), self.html)
+
+    def test_large_jinja_template_is_compiled_once(self):
+        with self.app.test_request_context("/plugins/A_pwmenu/"):
+            with mock.patch.object(
+                self.plugin,
+                "_get_html",
+                return_value="PWMenu {{ marker }}",
+            ) as get_html:
+                self.assertEqual(
+                    self.plugin._render_cached_template(marker="first"),
+                    "PWMenu first",
+                )
+                self.assertEqual(
+                    self.plugin._render_cached_template(marker="second"),
+                    "PWMenu second",
+                )
+
+        get_html.assert_called_once_with()
 
 
 if __name__ == "__main__":